Draft simple BiDi connection for Firefox nightly
