+++ rb/lib/selenium/webdriver.rb
@@ -36,6 +36,7 @@ module Selenium
     Rectangle = Struct.new(:x, :y, :width, :height)
     Location  = Struct.new(:latitude, :longitude, :altitude)
 
+    autoload :BiDi,       'selenium/webdriver/bidi'
     autoload :Chrome,     'selenium/webdriver/chrome'
     autoload :DevTools,   'selenium/webdriver/devtools'
     autoload :Edge,       'selenium/webdriver/edge'
